pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "frads"
dynamic = ["version"]
readme = "README.md"
authors = [
    {name = "LBNL", email = "taoningwang@lbl.gov"},
]
description = "Framework for lighting and energy simulations"
dependencies = [
    "epmodel==0.8.0",
    "numpy>=1.24.4",
    "pyradiance == 1.0.2",
    "pywincalc>=3.1.0",
    "pyenergyplus_lbnl>=23.2.0",
    "scipy>=1.10.1",
]
license = {"file"= "license.txt"}
requires-python = ">= 3.10"

[tool.setuptools.dynamic]
version = {attr = "frads.__version__"}

[tool.ruff.lint.pydocstyle]
convention = "google"
